Fix for merb.url in merb console causing the error NameError: undefined local variable or method `controller_name'.
The irb rack adapter url() method was attempting to set a fallback in
the call to Merb::Router.generate. The fallback was trying to use
controller_name and action_name which are not available in the
irb rack adapter. The explicit fallback has been removed and now the
call to Merb::Router.generate uses the default fallback.