Added support for dynamic layouts.
When a layout is specified as a symbol on the controller level, we now try to call
a method with that name to dynamically determine the layout. We fall back to
interpreting the symbol as the name of the layout if no method with that name exists.
This is mainly to support rendering a different layout based on request parameters,
for example mobile vs. web. See ticket #66.

Signed-off-by: Michael D. Ivey <ivey@gweezlebur.com>