Added Merb.testing? method
This will check for either $TESTING (as used by merb internals/plugins) or Merb::Config[:testing], which can be set in the config options. Having an option like this provides a nice interface for the more public facing parts of Merb; the auto-generated spec_helper.rb file for example can just use the test environment setting.