Added --sandbox (-S) option for IRB console
When specified, the console will call open_sandbox! which in turn calls any
ORM module's method of the same name (the modules Merb::Orm::* are
considered). It's up to the ORM implementation to start a transaction. When
you exit the IRB session, close_sandbox! will rollback any open transactions,
again by calling the appropriate ORM implementation of that method. See
Merb::Orms::ActiveRecord for an example.