Refactored Merb::AbstractController.layout class method
The approach is a generalization of how render options can be set from the
controller class(methods). Merb::AbstractController.render_options has been
introduced for this. The refactored implementation of the layout method sets
a :layout value in the default_render_options Hash. This class-level Hash is
merged at render time. The layout method now accepts false as well as nil as
arguments to indicate no layout should be used.