Fixed lurking infinite loop with not so common formats and ExceptionController
When a controller rendered a somewhat exotic output format (:css, :jpg ...) and
an exception was raised (non HTTP related, like ArgumentError) an infinite
would be triggered unless the ExceptionController provides that format. Which
isn't and shouldn't always be the case. The NotAcceptable exception that followed
would be retried/dispatched/rescued over and over.