Append the content_type to the given :template option for render()
With this change doing render :template => 'shared/foo' will try to locate
shared/foo.<content_type> first, and if that didn't return any results it
falls back to the old behaviour. As usual _template_location is called, but
in opts[:template] is given as the template context. The method signature
has been changed to fit the terminology better: an action name or :template
option can set the context for constructing a template location path. In the
latter case the controller argument is nil, to distinguish between the two.