Correct lable and line number in Router#compile
This patch fixes a bug in the function that compiles the Router#match function.  before this
change, any exceptions thrown by the generated code would claim to come from the
lib/merb-core/dispatch/router.rb file on a random line number > 87.  This lead to developers
who have errors in their project's route.rb looking in the wron place to discover why their
code broke.

This patch adds information to the eval statment that creates the Route#match function.  Error
messages will now discribe the "file" as generated code and accuratly point to the place it was
generated.  It also fixes the line number to match the optional debug output.

All Tests Pass.

	modified:   lib/merb-core/dispatch/router.rb

Signed-off-by: Michael S. Klishin <michael@novemberain.com>