Made HTTP method override proc-based and pluggable. [#364]
Changed the check for specific params to procs that execute
in the request scope, allowing checks for method overrides to
be based on params, headers -- or pretty much anything in the
request. Also changed the name of the array that stores these
checks from browser_method_workarounds to the more accurate
and descriptive http_method_overrides.

Signed-off-by: Michael D. Ivey <ivey@gweezlebur.com>
 (I changed this to yield the controller to the block, instead of
  using instance_eval, as it seemed cleaner and was significantly
  faster.  See benchmark attachment on ticket #364 --ivey)

[#364 state:resolved]