Added stacking and redirecting deferred routing blocks.
Added the ability to redirect requests from inside a deferred
block using the #redirect helper. Also made the #url helper
available in redirect blocks.

Defer blocks are now stackable using #defer. #defer_to is
still available, but it is a terminator (as in the route
cannot be further customized after.)

Also cleaned up a bunch of specs and miscellaneous routing
code.