Added the ability to specify resource actions in the block
Custom resource actions can now be specified in the block passed to
Behavior#resources and Behavior#resource. Doing it this way allows
the action to be customized further than by specifying it in in the
options hash. The API is as follows:

Merb::Router.prepare do
  resources :users do
    collection :boom, :method => :get, :to => "boom_get"
    collection :boom, :method => :put, :to => "boom_put"
    member     :boom, :method => :put, :to => "boom_put"
  end
end

Also added the ability to customize resource actions on singular
resource routes. This works the same way as the plural method,
except that only #member is available.